Opt out of Google's FLoC
CFPB should be a trusted resource for all of our visitors. They should be able to able to trust they won't be placed into a cohort to have, for example, high-interest loans advertised at them if they use our mortgage information and tools.

This change sets the `Permissions-Policy: interest-cohort=()` to declare consumerfinance.gov should not be included in a browser's cohort calculation, per [web.dev's FLoC explainer](https://web.dev/floc/#how-can-websites-opt-out-of-the-floc-computation).