Fixes DF-1935 - slick carousel error
Fixes DF-1935 - slick carousel error. Slick carousel was checking for
`slickObj` on all pages, so was throwing an error on everything but the
`/the-bureau/bureau-structure/` page. This adds conditional logic to
check for the required object before continuing with the slick carousel.
Also, removes slick carousel from the browser shims as it doesn’t need
shimming and sets up proper class export for ContentSlider module.