Changing H1 above intro paragraph
The existing text, "Our mission," is not 100% accurate and ended up confusing people in testing. After some discussion, we decided to change.

"The Bureau" was also under consideration, as it mirrors the sidebar text. But the video hero at the top made putting the title there seem a bit anticlimactic. And it felt like that intro paragraph would need a bit of rework. "Who we are" seems like we can leave that text as-is.