Fix mobile text intro (#1884)
Re-wrote the way we check for blank authors/links. It's clunky, but in a perfect world people would delete the blank author and link objects in the CMS and we don't live in that perfect world. :cry: Now we should be able to avoid having lots of blank mobile links when a content editor leaves a blank field.

* Remove empty list_link on mobile
* Rewriting how we check for blank authors/links