Adds skip link to on-demand header (#1923)
This PR adds the necessary CSS and JS to support skip links. Since this
header is going on different pages with unique page layouts we can't
hook into #main the same way we do on v1 pages. Instead there's a blank
div at the end of the navigation that will allow skip-nav users to skip
the navigation.

* Also silences breaking wagtail test. BEWDs should investigate.