Fixed issue with spacing after the last-child
The full-width text molecule always lives within a block, which has 60px of bottom spacing. The addition of bottom spacing to the last paragraph of the full-width text molecule was being stacked with the block’s 60px increasing the space more than spec’d. Removing the bottom margin from the paragraphs when they are a last-child fixes the issue.