Rebuilt Nemo Grunt tasks in Gulp
The Nemo styles and scripts were moved to the project but have no way of being updated. Created gulp tasks for the Nemo styles and scripts to make it possible to make changes and rebuild the assets.

NOTE: The assets were being built in the same directory as the original files. Updated the tasks to build them in the ignored static_built directory and removed the built files from the origination directories. Built files are now git ignored and must be processed on deployment.