Image compression (#2178)
* Image compression

This commit runs all of the images through the ImageOptim lossless
compressor. For the images that are over 100KB, it runs them through the
lossy compressor.

* Uncompressed Bureau Structure Chart

Since the Bureau Structure Chart is always downloaded and it's mostly
text, it makes sense to leave it uncompressed for maximum quality. I
noticed the headings were just a tad bit blurry.