fix tests by removing test-only models
Unfortunately Django doesn't make it easy to use test-only models, so
instead of simply doing this, I have to use some existing real model.
Here I chose to use django.contrib.auth.models as it's a safe bet that
it exists, but it's still not ideal to have this external dependency.

See this Django issue for more info on why this is needed:

https://github.com/django/django/pull/4700/