Add deployment smoke-test scripts
These scripts are intended to make sure the basics are working
immediately after a deployment. The scripts make requests that bypass
Akamai to check that pages and static assets are returning `200`
status codes.

This refactors the http_smoke_test script and moves it out of our
Jenkins-as-code repo. So the JAC job will need to be adjusted
to get the script from the refresh repo at /cfgov/scripts.

This PR also updates our testing docs to describe using `tox -e fast`
and how to see the coverage report.