This is an exploration in self-healing URLs
The goal is a URL that self-redirects to a page's current slug if it
changes or if something pollutes the request slug. The scheme avoids
adding a URL node to the Wagtail tree by depending solely on the slug
field. Each ask_cfpb slug gets a suffix that specifies language and
answer ID. That allows us to consult the specific page, check that
the slug matches, and either serve the page or redirect to the page's
current slug.