Adds creation of parent pages and reconfigures admin
To avoid having to repeatedly create parent pages during testing, this
adds their creation to the migration script, if they don't exist.
There are three parent answers, two for Spanish Answers and one for
English.

Also included are improvements to the organization and help text for the
Answer model's editing page in the admin.