fix secondary navigation page highlighting
This commit fixes a bug with how pages in sidebar navigation were not
being highlighted after a transition from HTTP to HTTPS. Previously the
logic to determine if a page was selected used

request.url.startswith(item.url)

which was prone to problems if the scheme of the two requests differed
(HTTP vs HTTPS, which could happen due to load balancers, etc). This
change adds a new item.active parameter which indicates whether or not
the page should be highlighted.

This code is fairly convoluted which makes it difficult to test
efficiently. The sharing logic is due to be removed soon so this will be
a quick fix for now.