always use root relative URLs in secondary nav
Instead of using v1.templatetags.share.get_page_state_url to determine
what URL to link to in secondary navigation sidebars, always use a
page's relative URL to the current request.site. This can be simplified
for Wagtail installations with a single Site, which would be able to
just use page.url.