Add json methods to Category and fix migration script
To implement in-browser faceting for category pages, we need to add json
methods to the Category model to deliver some small json mapping files
for slicing answers by audience and subcategory.

We also need to update the migration script to allow the Answer object
to do the work of adding feedback modules to an Answer page, so that
when pages are created in the future (i.e. not via migration), they will
have feedback modules.