remove obsolete custom link handler
This commit removes a custom rich text link handler (registered against
the `register_rich_text_link_handler` hook) that is obsolete after the
move to wagtail-sharing.

Compare this code against the core Wagtail
`wagtail.wagtailcore.rich_text.PageLinkHander` class:

https://github.com/wagtail/wagtail/blob/v1.8.1/wagtail/wagtailcore/rich_text.py#L21

Our custom code was doing a `urlsplit(page.url)` which will fail if the
page URL is relative or `None`. These cases would never happen
pre-wagtail-sharing because there were always multiple Wagtail Sites
defined, but now (and in standard Wagtail usage) this cannot be assumed
to be the case.

This restores the default Wagtail behavior of inserting `<a href="None">`
for links that point to pages that are not part of any Site root.