CFGOVPage.url always returns relative URLs
This commit overrides the standard `Page.url` behavior for the
`CFGOVPage` class so that it always returns a relative URL. This is to
enforce use of relative URLs in links when running Wagtail in a
multi-`Site` setup that shares a site root.

See comment in `v1.models.base.CFGOVPage.url` for more details.