Store rich text table cells like rich text fields.
The `v1.atomic_elements.organisms.AtomicTableBlock` block provides a
rich text editor in the Wagtail admin, but doesn't currently store its
content in the database in the same way that the built-in
`RichTextBlock` block type does. This causes some inconsistency, for
example in how internal links are rendered.

This change modifies the way that table content is stored in the
database, making sure that it is first cleaned using `DbWhitelister` and
then restored using `expand_db_html`, consistently with how Wagtail does
it for `RichTextBlock` content. See `wagtail.wagtailcore.richtext` for
reference.

This change also renames and slightly refactors the `AtomicTableInput`
to `RichTextTableInput` to be more explicit about how it is intended to
work. The JavaScript `initAtomicTable` method is similarly renamed to
`initRichTextTable`.