stash: render housing counselor results on server
This commit is a prototype approach to rendering HUD housing counselors
server-side instead of via AJAX. It's not ready for release yet, but
just demonstrates how this could be done.

There's a new `signed_redirect` Django template filter that can be used
to convert an external URL to a signed redirect one. This is now used in
the housing counselor template to modify external links.

The rest of the housing counselor content still needs to be added to the
page; in addition, I changed various HTML tags to have `-nojs` just to
prevent the JavaScript code from messing with them. Really that JS code
should be redacted to its absolute minimum (which might just be the map,
potentially).

To test this change you'll need to have the optional django-hud
application installed. Visit these URLs to see how it works with no
results, with a bad input, and with a valid input.

http://localhost:8000/find-a-housing-counselor/
http://localhost:8000/find-a-housing-counselor/?zipcode=foo
http://localhost:8000/find-a-housing-counselor/?zipcode=12345

There's still significant work to be done here but this is a first step.