Move legacy JS and CSS out of HUD API app
When we ported the templates from the find-a-counselor tool into
cfgov-refresh we left the static assets in the django-hud repo. This
is a poor separation of concerns because the tool's front-end is split
between two different Django apps. While updating the JS to support
server-side rendering I moved the static assets into cfgov-refresh's
legacy directory.

https://[GHE]/CFGOV/platform/issues/791#issuecomment-115539