Remove URL expansion in printed pages
We show URLs next to <a> tags when users print the HUD counselor
results. This is a common technique but in our case it's unnecessary
because we manually list every counselor's URL in the search results.
Coupled with the fact that most of the URLs are long `/external-site`
links makes it particularly confusing and unnecessary.

https://[GHE]/CFGOV/platform/issues/791