remove RuntimeWarning caused by test_data script
The `scripts.test_data` script is assigning a page tag that's type
str, not unicode, which causes the following RuntimeWarning:

/Users/chosaka/.virtualenvs/cfgov-refresh/lib/python2.7/site-packages/unidecode/__init__.py:46:
RuntimeWarning: Argument <type 'str'> is not an unicode object. Passing
an encoded string will likely have unexpected results.
  _warn_if_not_unicode(string)

This commit merely forces the test tag to be a unicode string.