Fix info unit groups' heading spacing
CF's sibling heading rules (e.g. h2 + h4) don't work with info unit
groups because the h2 is outside the content block so we add top margin
to headings that are inside info groups' content div. All of the info
group headings are optional so we also add margin to paragraphs.

https://[GHE]/CFGOV/platform/issues/860