better template tag for parsing/displaying services
The django-hud API returns available services as a comma-delimited
string with a custom hex code for the comma character. This commit
introduces a new `split_services` Django template tag that converts a
comma-delimited string like this into a simple list with proper commas.

This avoids needing to use `|safe` in the template.