Update the FCM to CFv4 markup and styles
The project is currently using a really old version of the Featured Content Module. Oddly, it was never updated to use the atomic version within this code base. This change updates the markup to the atomic pattern and duplicates the Less from the latest version in Capital Framework. The /organisms/featured-content-module.less can be removed when this project is updated to use CFv4.