Update the FCM to CFv4 markup and styles (#2939)
* Update the FCM to CFv4 markup and styles

The project is currently using a really old version of the Featured Content Module. This change updates the markup to the atomic pattern and duplicates the Less from the latest version in Capital Framework with a few improvements that are now open for review at https://github.com/cfpb/capital-framework/pull/493. The /organisms/featured-content-module.less file can be removed when this project is updated to use CFv4.