remove mysql from base requirements
This commit removes the MySQL-python requirement from the base Python
requirements and instead moves it to a new requirements/mysql.txt, which
is then included in requirements/test.txt and a new
requirements/production.txt.

This change allows for faster running of tests both locally and on
Travis by avoiding the need to install MySQL-specific requirements. It
also allows for running the code on machines that don't have MySQL
installed.

This change requires a related change in any build/deploy code that may
be depending on MySQL-python being include in requirements/base.txt.