Remove the negative margin on the FCM
The negative margin was causing the FCM to be pulled beyond the block container, which in turn caused the left hand border to be clipped on right sidebar templates. Removing the negative margin fixes this, but causes a slightly different issue of a double border on right sidebar templates. This turns out to be a rounding error, and is not something worth diving into at this exact moment due to discussion to remove the FCM to sidebar abutment.