generate housing counselor json and pdfs in bulk
This commit adds two new Django management commands to the `legacy` app:
`hud_generate_json` and `hud_generate_html`. These commands are intended
to support bulk generation of JSON and PDF files for the website's "find
a housing counselor" tool.

The `hud_generate_json` command downloads HUD counselor data from
data.hud.gov, cleans it, and sticks it into an in-memory SQLite
database, indexed by location (latitude/longitude). An external
reference of zipcode -> lat/lng pairs (downloaded from the Census.gov
website) is used to then lookup the closest counselors for every zipcode
in the country. A JSON blob is generated for each zipcode and dumped to
a target directory, with filenames like 20001.json. Command usage is
like:

manage.py hud_generate_json /path/to/zipcode/reference.txt target_dir

The `hud_generate_html` command then can be used to process that
directory full of JSON files to generate self-contained HTML files to be
fed into a PDF encoder. "Self-contained" here means that all CSS is
included inline and also that the single CFPB logo contained within is
embedded as a data URI. This is to avoid having to make multiple
requests when encoding as PDF. A new
`housing_counselor_pdf_selfcontained.html` template defines the HTML
used by this command. Filenames are generated like 20001.html, and
command usage is like:

manage.py hud_generate_html input_dir taget_dir

The JSON generated and used by this change varies slightly from that
generated by the existing django-hud app. Here, the counselor services
and languages are included in the JSON as a JSON array, instead of being
converted to a comma-separated list.

The data cleaning and standardization behavior introduced here is
primarily copied from the `load_hud_data` management command in the
django-hud project.

This code also includes a simple geocoder (currently commented out) that
has the capability to add missing latitude/longitude information for
counselors that are missing it; this just uses the zipcode centroid to
avoid making an external geocoder call. This copies some legacy behavior
from the django-hud project. Because of the way we query the HUD API,
this is actually never needed, but the code remains should it be useful
in future.