don't let sheerlike replace the context's request
The default SheerlikeContext used by Jinja tries to use a local thread
as a place to stash the request, and then picks it back up during
rendering. This can cause problems with unit tests because an old
request may be passed to a later rendering of a template, for example if
testing rendering a request object that hasn't gone through the
SheerlikeMiddleware.

This change just prevents the SheerlikeContext from replacing any
existing request object.