Fix Ask CFPB's SubCategory admin and indexing
The admin has an errant field ("featured") that applies to answers, not
to subcategories. This causes a 500 error when trying to use the Wagtail
admin to create a new subcategory.

This PR also removes the AnswerTagProxy model and related code, as it is
no longer needed. We also resurrect the knowledgebase templatetag for
stripping accents, which IS needed.