Ask CFPB: Fix server request errors
Server errors have cropped on pages like these:
- /es/obtener-respuestas/buscar/?q=cooperativa de credito&page=6
- /es/obtener-respuestas/buscar-por-etiqueta/dinero/?page=10
- /es/obtener-respuestas/categoria-prestamos-de-dia-de-pago/None/
- /es/obtener-respuestas/que-es-una-empresa-de-informes-de-credito-es-1251/

The first three are empty-page errors, which the English pages are
vulnerable to as well, and the last is a feedback
encoding error in which the referring page's URL has unicode characters.

The Spanish pages seem to be getting hit by malformed URLs that wouldn't
be created by our site, but this PR should catch them.