Add the sortable table options to the Wagtail admin
This is step one of adding sortable tables to Wagtail. It includes the admin options, the front-end markup, and a feature flag to protect the front-end. The front-end won't work until the project has been updated to use cf-tables v4.x. When that work is complete, the feature flag should be removed.