fix handling of unicode table block headers
If TableBlock header values contain unicode characters, the logic that
sticks the header value into the cell "data-label" attribute raises a
UnicodeEncodeError.

This commit fixes that handling so that unicode content is properly
placed in the page, and adds a regression test to cover this behavior.

To reproduce, create a TableBlock and check "Stacked on mobile" and
"Row header". Then enter some unicode characters into a column header
and set some value in the next row. Previewing the page (or viewing a
published version) should raise the exception on master but not with
this fix.