date must be provided if used in RelatedMetadata
The RelatedMetadata molecule provides an option for a date entry which
consists of a heading field and a date field. This change makes the date
field required (previously it was optional). The template:

/jinja2/v1/_includes/molecules/related-metadata.html

assumes that the date is filled in. If left empty, rendering of the
element will raise an exception and cause a 500 error. This change makes
it so that Wagtail will fail form validation if users try to enter a
date element in RelatedMetadata without specifying the date.