add AWS_STORAGE_BUCKET_NAME to .env_SAMPLE
As of #3107, housing counselor JSON and PDF data will be served on the
website from S3. Currently this is behind a feature flag but will be
made default soon.

This change requires knowing which S3 bucket the data lives on, which
must be defined using the `AWS_STORAGE_BUCKET_NAME` environment
variable. This change adds that variable (with value
`files.consumerfinance.gov`) to the sample `.env_SAMPLE` file.