Ask CFPB: Fix Answer page previews for admin
Wagtail-sharing requests for Answer preview pages are not routing
through the sharing framework because Answer pages are routed through a
self-healing view outside of Wagtail. The view now checks the request to
see if it is coming from a sharing site, and if so renders a preview.

This addresses GHE issue `#1020``