qAdd data_research API for mortgage performance data.
The goal of this (initially internal) API is to produce charts and other
visualizations based on nationwide county-level mortgage status data,
extrapolated to metro areas, states and the nation.

Data will be updated quarterly, and this PR is the first step
in building a data pipeline for automating those updates, and a download
delivery scheme to meet our mandate to release national mortgage data.