deprecate HOUSING_COUNSELOR_S3 flag
This change removes the `HOUSING_COUNSELOR_S3` feature flag and makes
default the behavior that it was guarding. This modifies how the "Find a
housing counselor page" (at `/find-a-housing-counselor/`) serves JSON
and PDF data, always serving it from S3.

This change leaves intact the existing API endpoints for rendering JSON
and PDF data on the fly from django-hud database tables
(`hud-api-replace/18020/` for JSON,
`/find-a-housing-counselor/?zipcode=18020&pdf` for PDF-able HTML, and
`/save-hud-counselors-list/?zip=18020` for PDFs).