properly include page streamblock children media
The `CFGOVPage.media` property returns a dictionary of JS files that
were included by child blocks on the page. This logic had a slight bug
(discovered by @jimmynotjim) where media files defined on blocks that
were nested inside of a `StreamBlock` (or block descendant thereof) were
not being included.

This change modifies the `CFGOVPage.media` logic to properly include
`StreamBlock` children media in the same way that `StructBlock`
children media is included. It also adds a couple of unit tests to
validate this behavior.

There's no easy way to replicate this right now without modifying an
existing block to define a `class Media`; one example would be
`v1.atomic_elements.organisms.AtomicTableBlock`. If you add a media file
there, you'll see that current Wagtail behavior includes that media file
if the table block is added as a top-level block under a content
`StreamField`, but not if added within a nested `FullWidthText` in that
content. This change fixes that behavior.