roll back model creation in unit test
This commit rolls back a commit in #3220

(bd44933e1c61f52681b0affa9af29cd979b515590)

that tried to create a temporary model for use in a unit test. This may
be causing some problems with running the unit tests in isolation.

This change rewrites those tests to unfortunately be less comprehensive
but just use existing models to test the functionality.