Create elements for mortgage performance interactive
This creates a block and page to satisfy requirements of displaying
visualizations for the mortgage performance project. The chart block
created for consumer credit trends, which handles single charts with set
parameters, cannot be reasonably exntended to serve the needs of the
mortgage visualizations, which offer users many geographic choices
for making comparisons. There are also more visualizations planned that will
benefit from context values delivered via the new page.