Mortgage performance: add aria-label attributes
Our downloads page has visual clues to distinguish download links, but
the link text is the same, which doesn't help people using screen
readers. This adds aria-label attributes to download links with
specifics about what is being downloaded.

Changes are not visible but can be checked in dev tools on the [download
page](http://localhost:8000/data-research/mortgage-performance-trends/download-the-data/)