Move MP map store into constructor
The map's starting date needs to be set to whatever month/year data
is most recently available for. This value is passed to the JS via a
data-attribute set in the Jinja template. In order to initialize the
app's store with the correct date we have to wait until the data
attribute is read from the DOM before creating the store.