add request where available to github error issues
The `alerts.logging_handlers.CFGovErrorHandler` logging handler posts
logging messages as issues to the GitHub API. Currently this issue gets
its title from the log message and its body from the exception attached
to the message.

This information can sometimes be insufficient to debug an error, for
example when query string parameters cause a problem. These parameters
aren't necessarily included in either the log message or the exception
stack trace.

Django's default error emails (in `django.utils.log.AdminEmailHandler`)
include a representation of the request. This includes the full path
being requested, the GET parameters, cookie information, etc. This
change modifies `CFGovErrorHandler` to duplicate the content of Django
500 error emails, and use that content as the GH issue body.

Unfortunately due to how the Django 1.8 code is written, this requires
some duplication with Django code. As of Django 1.9+ this could probably
be rewritten to be a little bit nicer.

A note about privacy concerns: this change does add additional
information to the GH issues that are created, but, again, this is
consistent with information that was previously in error emails. Django
does provide the ability to filter the data (for example, excluding
sensitive POST parameters), and this change invokes the code that would
do that, if configured. See the Django docs for more on this:

https://docs.djangoproject.com/en/1.8/howto/error-reporting/#filtering-error-reports

Pinging @ascott1 @schaferjh for any comments on privacy considerations.

Finally, this change modifies the existing unit tests for this code, to
test the end-to-end `logging.error` functionality instead of the handler
itself. This seemed more consistent with how Django tests its own
handler and also requires less manual creation of e.g. `LogRecord`
objects.