Clean up form submission handler in test
JSDOM does not allow for a form to be submitted and will throw an error.
To avoid any issues we've prevented it from submitting. This change
relocates that code outside of the button click handler and expands on
the comment to make it clearer they are unrelated and that the
submission handler is an outside necessity.