only include streamfield media for existing blocks
This change fixes some logic that was trying to include StreamField
child block media. We want pages to include all child block media
(block-specific JavaScript files) for any blocks that are on the page,
but the existing logic includes any files for any blocks that could
possibly be on the page.

This change leverages some logic in wagtail-inventory that lists the
block types on a page, instead of incorrectly pulling in all possible
child blocks of a parent block.

A few unit tests had to be changed because using model_mommy.prepare on
Wagtail page types doesn't properly set the Django contenttypes id,
which breaks use of page.specific.

This addresses the first issue in GHE#1264.