flatten CFGOVPage.media from dict to list
This change flattens the @media property on CFGOVPage to return a simple
list of JavaScript filenames instead of a dictionary. The dictionary
keys were never used downstream in the template, so maintaining them
serves no purpose and makes it harder to test for files by checking for
an empty dictionary.

This change also slightly cleans up the class methods around this
functionality, converting CFGOVPage.add_page_js(js) (which adds to a
dictionary) into CFGOVPage.page_js (which just returns a list).

Several AskCFPB classes and tests had to be modified for this; pinging
@higs4281 for a quick review.

Note that this change maintains the existing requirement that page/block
JS filenames have to be globally unique.