Match-all-tags filtering
Editors helped us discover a combination of page tags and the use of
'match-all-topic tags' that locked up the version of mysql that we run
on servers. This PR reduces some of the query complexity and, in the case
of 'match-all,' forces a query evaluation before starting the recursive
tag filtering routine.

I replicated the mysql lockup by editing page #10848 inside a docker vm
with the same db version used in production, then confirmed that
this commit stopped the lockup.

When the final query succeeded, applying 'match-all' to all 9 tags,
the result was zero related pages.
All this trouble for a null!