remove print() in agreements tests
This change removes a `print` statement that was getting called in unit
tests. As of #3381 the agreements app was moved into cfgov-refresh, but
one of those unit tests included a `print` statement.

This change modifies those tests so that the management command instead
does the logging, and respects its `verbosity` option to enable/disable
the logging.

A new unit test has been added to verify that the print statement is
properly called by the command.