mattermost alert failures shouldn't be fatal
This change modifies how Mattermost alert failures are handled as part
of the SQS alert consumption process.

Currently, when an alert is read off of an incoming SQS queue, the
process goes like this:

1. Peek at alert on top of queue.
2. Post alert to GH.
3. Post alert to MM.
4. Delete alert from queue.

If Mattermost happens to be down (as happened recently) and an alert
comes it, it'll be repeatedly posted over and over to GH, because a
failure at step 3 prevents alert cleanup at step 4.

This change catches exceptions from MM posts and logs them, instead of
allowing the alert to exit the process.

In practice that means that if this script is run from e.g. Jenkins the
Jenkins job wil not fail just because MM is down.