log filtering messages as debug, not info
This change modifies the logging in two places where pages get filtered.
Right now these filters get logged at the INFO level, which means they
go to the console and log files by default.

This fills up the logs with filtering every time a filter is chosen,
which seems overly verbose. If we want these to be logged, could we
instead log them to DEBUG?