remove deprecated staging middleware
This commit removes the deprecated `v1.middleware.StagingMiddleware`
which was needed in an old implementation of page staging before the
integration of wagtail-sharing.

The middleware added an `is_staging` boolean onto requests that were to
the staging site; this flag is not used anywhere and so is removed.

The unit tests for this middleware are also removed, as is the Django
setting (`STAGING_HOSTNAME`) that powered it.