dynamically populate snippetlist choices
This change fixes a bug that was introduced by the module reordering in
PR #3602. Because the SnippetList organism was defining its choices at
import time, it relied on snippet classes having been imported first.
This cannot be guaranteed.

This change modifies how SnippetList populates its choices, using
callables that get executed at runtime.

A new unit test verifies that some snippet choices get added; this can
also be tested in the Wagtail admin by adding a SnippetList organism and
confirming that the choices are in the dropdown.