even simpler approach to flagged drafts
This change is a follow up / alternative to #3707 to support exposure of
draft content given certain conditions. Instead of relying on a
hardcoded set of page primary keys, it utilizes built-in wagtail-flags
capability to control draft exposure via feature flag.

So, for example, Wagtail users could create a flag for a particular site
with the condition:

"path matches": "/path/to/page/"

and that page will be served with its draft content. A more complicated
regular expression would support multiple pages like this, e.g.

"patch matches": "/page1/|/page2/"

This lets draft serving be controlled through the Wagtail UI, instead of
relying on code changes.