add second email popup for debt collection pages
This change modifies the existing JS email popup so that it can appear
on a wider set of pages, and also supports multiple simultaneous popups
at the same time with different content and target email lists.

It leaves the existing Owning a Home popup on these pages:

- /owning-a-home/
- /owning-a-home/mortgage-estimate/

and adds a new Debt Collection popup on these pages:

- /ask-cfpb/what-is-a-statute-of-limitations-on-a-debt-en-1389/
- /ask-cfpb/what-is-the-best-way-to-negotiate-a-settlement-with-a-debt-collector-en-1447/
- /ask-cfpb/what-should-i-do-when-a-debt-collector-contacts-me-en-1695/
- /consumer-tools/debt-collection/

This new popup will not turn on until February 1st, and can be
configured using the EMAIL_POPUP_DEBT feature flag. The existing
EMAIL_POPUP feature flag that was previously being used to control the
OaH popup has now been renamed to EMAIL_POPUP_OAH, and is enabled by
default.

Specific popup content and configuration, including text/image content
and the mailing list to subscribe users to, can be configured using the
Django template files in

cfgov/jinja2/v1/_includes/organisms/email-popup/

for example, oah.html and debt.html. This use of code-based
configuration instead of full integration in the Wagtail editor is
deliberate.

A new "email_popup" template tag has been created that works in both
Django and Jinja templates. This tag inserts the HTML code for the email
popup if the page request is configured for a popup in settings. The
necessary JS code for each page must also be setup through the use of
cfgov/unprocessed/js/routes files.

The existing email popup uses two browser local storage keys to store
information about how many times the user has seen the popup, and when
they should see it next. These keys are used to throttle how often the
popup appears.

The current OaH popup uses the keys oahPopupCount and oahPopupShowNext.
The new debt popup uses the keys debtPopupCount and debtPopupShowNext.
For debugging/testing purposes these keys can be viewed and deleted in
Chrome Developer Tools under Application, Storage, Local Storage, and
then the domain of the site you are testing against (e.g. localhost).

The best way to test this change (besides the unit tests) is to run a
local server and visit the pages listed above to see if the popup
appears. In order to enable the debt popup before the go-live date of
2/1, you'll need to manually set the EMAIL_POPUP_DEBT flag to something
via the Wagtail admin, for example a boolean condition of "True".

Finally note that the use of modal-house.png for the debt popup is a
temporary placeholder pending new imagery.